+++ lib/widget_tree.dart
@@ -2,49 +2,108 @@ part of '../cake_flutter.dart';
 
 class WidgetTree extends _WidgetTree {
   final Element _rootElement;
-  WidgetTree(this._rootElement);
+  WidgetTree(this._rootElement, {required super.indexOptions});
 
-  void index(WidgetTester tester) {
+  void index(IndexOptions indexOptions, WidgetTester tester) {
     elementWrapper = TestElementWrapper(_rootElement, tester);
-    _findChildren(tester);
+    if (indexOptions.matchesIndexFilter(_rootElement)) {
+      parentMatchesIndexFilter = true;
+    }
+    _WidgetTree._findChildren(
+      this,
+      _rootElement,
+      tester,
+    );
   }
 }
 
 class _WidgetTreeNode extends _WidgetTree {
   final Element nodeElement;
-  final int depth;
 
-  _WidgetTreeNode(this.nodeElement, {required this.depth});
+  _WidgetTreeNode(
+    this.nodeElement, {
+    required super.debugDepth,
+    super.parentMatchesDebugFilter,
+    required super.indexOptions,
+  }) : super(parentMatchesIndexFilter: true);
 
   void create(WidgetTester tester) {
     elementWrapper = TestElementWrapper(nodeElement, tester);
-    return _findChildren(tester, depth: depth + 1);
+
+    // Check if this node should be printed for debugging
+    if (indexOptions.matchesDebugFilter(nodeElement) ||
+        parentMatchesDebugFilter) {
+      String message = debugDepth > 0 ? '|' : '';
+      for (int i = 0; i < debugDepth; i++) {
+        message += ' ';
+      }
+      message += 'Found: ${nodeElement.widget.runtimeType}';
+      // ignore: avoid_print
+      print(message);
+      parentMatchesDebugFilter = true;
+    }
+
+    _WidgetTree._findChildren(
+      this,
+      nodeElement,
+      tester,
+      debugDepth: parentMatchesDebugFilter ? debugDepth + 1 : 0,
+    );
   }
 }
 
 abstract class _WidgetTree {
-  late TestElementWrapper elementWrapper;
+  TestElementWrapper? elementWrapper;
   final List<_WidgetTreeNode> _children = [];
+  bool parentMatchesIndexFilter;
+  bool parentMatchesDebugFilter;
+  final int debugDepth;
+  IndexOptions indexOptions;
 
-  _WidgetTree();
+  _WidgetTree({
+    required this.indexOptions,
+    this.debugDepth = 0,
+    this.parentMatchesIndexFilter = false,
+    this.parentMatchesDebugFilter = false,
+  });
 
-  void _findChildren(WidgetTester tester, {int depth = 0}) {
-    elementWrapper.element.visitChildren(
+  static void _findChildren(
+    _WidgetTree tree,
+    Element parentElement,
+    WidgetTester tester, {
+    int debugDepth = 0,
+  }) {
+    parentElement.visitChildren(
       (element) {
-        _add(element, tester, depth: depth);
+        _add(tree, element, tester, debugDepth: debugDepth);
       },
     );
   }
 
-  void _add(Element child, WidgetTester tester, {int depth = 0}) {
-    final _WidgetTreeNode tree = _WidgetTreeNode(child, depth: depth);
-    tree.create(tester);
-    _children.add(tree);
-    elementWrapper._addChild(tree.elementWrapper);
+  static void _add(
+    _WidgetTree tree,
+    Element child,
+    WidgetTester tester, {
+    int debugDepth = 0,
+  }) {
+    if (tree.parentMatchesIndexFilter ||
+        tree.indexOptions.matchesIndexFilter(child)) {
+      final _WidgetTreeNode node = _WidgetTreeNode(
+        child,
+        debugDepth: debugDepth,
+        parentMatchesDebugFilter: tree.parentMatchesDebugFilter,
+        indexOptions: tree.indexOptions,
+      );
+      node.create(tester);
+      tree._children.add(node);
+      tree.elementWrapper?._addChild(tree.elementWrapper!);
+    } else {
+      _findChildren(tree, child, tester);
+    }
   }
 
   TestElementWrapper? searchKey(Key searchCriteria) {
-    if (elementWrapper.key == searchCriteria) {
+    if (elementWrapper?.key == searchCriteria) {
       return elementWrapper;
     }
 
@@ -69,10 +128,15 @@ abstract class _WidgetTree {
   TestElementWrapperCollection<W> searchType<W extends Widget>([
     TestElementWrapperCollection<W>? collection,
   ]) {
+    // Ensure that this type was indexed in the first place
+    if (!indexOptions.hasIndexedType<W>()) {
+      throw '$W has not been indexed! Searching by type will fail without it.\nEither add WidgetType<$W>() to indexFilter or remove all filters.\n';
+    }
+
     collection ??= TestElementWrapperCollection<W>();
 
-    if (elementWrapper.widget is W) {
-      collection.add(elementWrapper.asType<W>());
+    if (elementWrapper?.widget is W) {
+      collection.add(elementWrapper!.asType<W>());
     }
 
     for (var element in _children) {
@@ -88,8 +152,8 @@ abstract class _WidgetTree {
   ]) {
     collection ??= TestElementWrapperCollection();
 
-    if (searchMatcher(elementWrapper)) {
-      collection.add(elementWrapper);
+    if (elementWrapper != null && searchMatcher(elementWrapper)) {
+      collection.add(elementWrapper!);
     }
     for (var element in _children) {
       element._search(searchMatcher, collection);