+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0 (2023-12-26)
+- [ADD] Added IndexOptions to search. IndexOptions currently allows for only certain widget types to be indexed for better search performance and shows debugging options for printing to the console.
+- [ADD][Search] Search will now throw an error if an un-indexed type is searched.
+
 ## 0.2.0 (2023-12-25)
 - [ADD][Expect] Added isWidgetType to match against a TestElementWrapper.
 - [ADD] Added a reference to children, if any, for each TestElementWrapper.